refactor: delete mod wrapper; wire DTOs live in message_pack_format
The `wrapper/` module was a parallel set of "wire-format-aligned"
sketch types added when the runtime sketches in `sketches/` were
believed not to produce Go-byte-compatible output. The parity probes
added in this branch (`tests/sketches_go_parity_probe.rs`) confirm
`sketches::Count<Vector2D<i64>, FastPath>`,
`sketches::CountMin<Vector2D<f64>, FastPath>`,
`sketches::HyperLogLogImpl<ErtlMLE, P14>`, and `sketches::DDSketch`
all emit byte-identical envelopes to `sketchlib-go` for the same
goldens, so the parallel layer is redundant.

Changes:
- Move every type formerly under `src/wrapper/*.rs` into the
  corresponding `src/message_pack_format/portable/*.rs` file, keeping
  the wire DTO and runtime ops together. Internal tests come along.
- Drop the `crate::wrapper` module entirely.
- Re-export the runtime + wire DTO types at the crate root from their
  new home in `message_pack_format::portable::*`.
- Add small public accessors to `sketches::DDSketch`
  (`alpha`/`sum`/`store_counts`/`store_offset`) and to
  `sketches::HyperLogLogImpl::registers_as_slice` so the parity probe
  can build wire envelopes directly from `sketches::*` without
  reaching into private fields.
- Extend `tests/sketches_go_parity_probe.rs` with HLL + DDSketch
  probes (CountSketch + CountMin already there); all 4 pass.
