+++ src/message_pack_format/portable/delta_set_aggregator.rs
@@ -1,7 +1,8 @@
-//! MessagePack wire-format for the delta set aggregator (see
-//! [`crate::wrapper::delta_set_aggregator`]).
+//! Wire-format DTO for the delta set aggregator.
 //!
 //! Owns the [`DeltaResult`] wire DTO and its [`MessagePackCodec`] impl.
+//! The streaming/state-tracking logic lives in the downstream consumer
+//! (ASAPQuery's accumulators) — only the over-the-wire shape lives here.
 
 use std::collections::HashSet;
 
@@ -9,8 +10,8 @@ use serde::{Deserialize, Serialize};
 
 use crate::message_pack_format::{Error as MsgPackError, MessagePackCodec};
 
-/// Wire DTO for the delta set aggregator. Public — re-exported through
-/// [`crate::wrapper::delta_set_aggregator`] for backwards compatibility.
+/// Wire DTO for the delta set aggregator: a snapshot of added/removed
+/// string keys between two consecutive observations.
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct DeltaResult {
     pub added: HashSet<String>,
@@ -24,5 +25,41 @@ impl MessagePackCodec for DeltaResult {
 
     fn from_msgpack(bytes: &[u8]) -> Result<Self, MsgPackError> {
         Ok(rmp_serde::from_slice(bytes)?)
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_msgpack_round_trip() {
+        let mut added = HashSet::new();
+        added.insert("web".to_string());
+        added.insert("api".to_string());
+
+        let mut removed = HashSet::new();
+        removed.insert("db".to_string());
+
+        let bytes = DeltaResult { added, removed }.to_msgpack().unwrap();
+        let result = DeltaResult::from_msgpack(&bytes).unwrap();
+
+        assert_eq!(result.added.len(), 2);
+        assert!(result.added.contains("web"));
+        assert!(result.added.contains("api"));
+        assert_eq!(result.removed.len(), 1);
+        assert!(result.removed.contains("db"));
+    }
+
+    #[test]
+    fn test_empty_sets() {
+        let dr = DeltaResult {
+            added: HashSet::new(),
+            removed: HashSet::new(),
+        };
+        let bytes = dr.to_msgpack().unwrap();
+        let result = DeltaResult::from_msgpack(&bytes).unwrap();
+        assert!(result.added.is_empty());
+        assert!(result.removed.is_empty());
     }
 }
