+++ tests/sketches_go_parity_probe.rs
@@ -8,11 +8,16 @@
 //! redundant.
 
 use asap_sketchlib::common::DataInput;
+use asap_sketchlib::common::hash::CANONICAL_HASH_SEED;
+use asap_sketchlib::common::structures::HllBucketListP14;
 use asap_sketchlib::proto::sketchlib::{
-    CountMinState, CountSketchState, CounterType, SketchEnvelope, sketch_envelope::SketchState,
+    CountMinState, CountSketchState, CounterType, DdSketchState, HyperLogLogState, SketchEnvelope,
+    sketch_envelope::SketchState,
 };
 use asap_sketchlib::sketches::countminsketch::CountMin;
 use asap_sketchlib::sketches::countsketch::Count;
+use asap_sketchlib::sketches::ddsketch::DDSketch;
+use asap_sketchlib::sketches::hll::{ErtlMLE, HyperLogLogImpl};
 use asap_sketchlib::{DefaultXxHasher, FastPath, Vector2D};
 use prost::Message;
 
@@ -27,6 +32,16 @@ const COUNTSKETCH_GOLDEN_HEX: &str = include_str!("cs_envelope_golden.hex");
 /// `(rows=4, cols=2048)` × `goldenCmsKeys()` (10 keys "flow-0".."flow-9",
 /// each repeated 5×). Same file the wrapper test already uses.
 const COUNTMIN_GOLDEN_HEX: &str = include_str!("../src/sketches/testdata/cms_envelope_golden.hex");
+
+/// 16 398-byte envelope captured from `sketchlib-go::HyperLogLog.SerializePortable`
+/// for precision=14, `(1..=50)` IEEE-754-LE byte-key input. Same file
+/// the wrapper test already uses.
+const HLL_GOLDEN_HEX: &str = include_str!("../src/sketches/testdata/hll_envelope_golden.hex");
+
+/// 432-byte DDSketch envelope captured from
+/// `sketchlib-go::DDSketch.SerializeProtoBytes` for `alpha=0.01`,
+/// `(1..=50)` integer-as-f64 input.
+const DDSKETCH_GOLDEN_HEX: &str = "080172ab03096214ae47e17a843f128003000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000100000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000100000000000000000000000000000100000000000000000000010000000000000000010000000000000001000000000001000000000001000000000001000000010000000001000000010000010000000100000100000100000100000100010000010001000100010001000100010001000100010100010100010100010101000101010100010101010101010100000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000187f2032290000000000ec934031000000000000f03f390000000000004940";
 
 #[test]
 fn sketches_count_fastpath_matches_go_count_sketch_envelope() {
@@ -141,6 +156,108 @@ fn sketches_countmin_fastpath_matches_go_count_min_envelope() {
     assert_eq!(
         got, want,
         "sketches::CountMin<Vector2D<f64>, FastPath> envelope diverges from Go golden",
+    );
+}
+
+#[test]
+fn sketches_hll_classic_matches_go_envelope() {
+    // Go's `HyperLogLog.SerializePortable` emits variant=DATAFUSION=2
+    // (ErtlMLE). Use the same variant + precision=14 + same input here.
+    let mut sk: HyperLogLogImpl<ErtlMLE, HllBucketListP14, DefaultXxHasher> =
+        <HyperLogLogImpl<ErtlMLE, HllBucketListP14, DefaultXxHasher>>::new();
+    for i in 1..=50i32 {
+        let v = i as f64;
+        let bytes = v.to_le_bytes();
+        let hashed = <DefaultXxHasher as asap_sketchlib::common::SketchHasher>::hash64_seeded(
+            CANONICAL_HASH_SEED,
+            &DataInput::Bytes(&bytes),
+        );
+        sk.insert_with_hash(hashed);
+    }
+
+    let state = HyperLogLogState {
+        // 2 = DATAFUSION on the Go wire (matches ErtlMLE)
+        variant: 2,
+        precision: 14,
+        registers: sk.registers_as_slice().to_vec(),
+        hip_kxq0: 0.0,
+        hip_kxq1: 0.0,
+        hip_est: 0.0,
+    };
+    let envelope = SketchEnvelope {
+        format_version: 1,
+        producer: None,
+        hash_spec: None,
+        sketch_state: Some(SketchState::Hll(state)),
+    };
+    let mut got = Vec::with_capacity(envelope.encoded_len());
+    envelope.encode(&mut got).expect("prost encode");
+
+    let want = decode_hex(HLL_GOLDEN_HEX);
+    assert_eq!(
+        got.len(),
+        want.len(),
+        "HLL envelope length: got {} want {}",
+        got.len(),
+        want.len(),
+    );
+    assert_eq!(
+        got, want,
+        "sketches::HyperLogLogImpl<ErtlMLE, P14> envelope diverges from Go golden",
+    );
+}
+
+#[test]
+fn sketches_ddsketch_matches_go_envelope() {
+    let mut sk = DDSketch::new(0.01);
+    for i in 1..=50i32 {
+        sk.add(&(i as f64));
+    }
+
+    let count = sk.get_count();
+    // Go serializes `alpha` round-tripped through gamma to mirror its
+    // internal storage: `alpha_wire = (γ-1)/(γ+1)`, `γ = (1+α)/(1-α)`.
+    // This is a 25-ULP shift from the user-supplied α=0.01.
+    let alpha = sk.alpha();
+    let gamma = (1.0 + alpha) / (1.0 - alpha);
+    let alpha_wire = (gamma - 1.0) / (gamma + 1.0);
+    let state = DdSketchState {
+        alpha: alpha_wire,
+        store_counts: sk.store_counts().to_vec(),
+        store_offset: sk.store_offset(),
+        count,
+        sum: sk.sum(),
+        min: if count == 0 {
+            f64::INFINITY
+        } else {
+            sk.min().unwrap_or(f64::INFINITY)
+        },
+        max: if count == 0 {
+            f64::NEG_INFINITY
+        } else {
+            sk.max().unwrap_or(f64::NEG_INFINITY)
+        },
+    };
+    let envelope = SketchEnvelope {
+        format_version: 1,
+        producer: None,
+        hash_spec: None,
+        sketch_state: Some(SketchState::Ddsketch(state)),
+    };
+    let mut got = Vec::with_capacity(envelope.encoded_len());
+    envelope.encode(&mut got).expect("prost encode");
+
+    let want = decode_hex(DDSKETCH_GOLDEN_HEX);
+    assert_eq!(
+        got.len(),
+        want.len(),
+        "DDSketch envelope length: got {} want {}",
+        got.len(),
+        want.len(),
+    );
+    assert_eq!(
+        got, want,
+        "sketches::DDSketch envelope diverges from Go golden",
     );
 }
